DL4J Issues/Fixes (#6634)
* #6611 Allow nets with more than Integer.MAX_VALUE params; fixes for indexing in large arrays

* Ignore annotation on new tests due to memory requirements

* #6609 Android compatibility for VersionCheck (remove use of SimpleFileVisitor)

* #6619 BarnesHutTsne.saveAsFile tweak

* #6632 Nd4j readTxt validation + extra tests

* Different fix for #6609

* Small stats listener tweak

* Small test fix

+++ deeplearning4j/deeplearning4j-core/src/test/java/org/deeplearning4j/optimize/solver/TestOptimizers.java
@@ -350,7 +350,7 @@ public class TestOptimizers extends BaseDL4JTest {
         }
 
         @Override
-        public int numParams(boolean backwards) {
+        public long numParams(boolean backwards) {
             return 0;
         }
 
@@ -433,7 +433,7 @@ public class TestOptimizers extends BaseDL4JTest {
             conf.addVariable("W"); //Normally done by ParamInitializers, but obviously that isn't done here
 
             Model m = new RastriginFunctionModel(10, conf);
-            int nParams = m.numParams();
+            int nParams = (int)m.numParams();
             if (i == 0) {
                 m.computeGradientAndScore(LayerWorkspaceMgr.noWorkspaces());
                 scores[0] = m.score(); //Before optimization
@@ -538,7 +538,7 @@ public class TestOptimizers extends BaseDL4JTest {
         }
 
         @Override
-        public int numParams(boolean backwards) {
+        public long numParams(boolean backwards) {
             return 0;
         }
 
@@ -738,7 +738,7 @@ public class TestOptimizers extends BaseDL4JTest {
         }
 
         @Override
-        public int numParams(boolean backwards) {
+        public long numParams(boolean backwards) {
             return 0;
         }
 
@@ -896,7 +896,7 @@ public class TestOptimizers extends BaseDL4JTest {
         }
 
         @Override
-        public int numParams() {
+        public long numParams() {
             // FIXME: int cast
             return (int) parameters.length();
         }
